requirements.txt
astroid==2.4.2
ImageHash==4.2.0
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.5
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
Pillow==8.1.0
pylint==2.6.0
pylsd==0.0.2
PyWavelets==1.1.1
scipy==1.6.0
Shapely==1.7.1
six==1.15.0
toml==0.10.2
wrapt==1.12.1
